// src/typings/model.ts
type bgStyles = {
  type: string;
  styles?: any;
  size?: number;
};
export type NodeIconConfigs = {
  /**
   * Percentage positioning of the icon relative to the node.
   * 图标相对于节点的百分比定位(0-1)
   */
  position: number[];
  /**
   * Customize the icon styles.
   * 自定义图标样式
   */
  setStyles: (nodeData: any) => { icon: string; size?: number; [k: string]: unknown };
  /**
   * The way to display the icon.
   * 图标展示方式
   */
  show?: 'always' | 'hover';
  /**
   * Absolute offsets of the icon(px).
   * 图标相对于节点的绝对偏移(px)
   */
  offsets?: number[];
  /**
   * Click handler of the icon.
   * 图标点击回调
   */
  onClick?: (e: ShapeEvent, nodeData: any) => void;
  /**
   * Mouse enter handler of the icon.
   * 图标移入锚点回调
   */
  onMouseEnter?: (e: ShapeEvent, nodeData: any) => void;
  /**
   * Mouse leave handler of the icon.
   * 图标移出锚点回调
   */
  onMouseLeave?: (e: ShapeEvent, nodeData: any) => void;
  /**
   * Customize the background of the icon.
   * 自定义图标背景样式
   */
  setBgStyles?: (nodeData: any) => bgStyles;
};
// ...
export type AnchorConfigs = {
  /**
   * The type of anchor, with the default being a circular dot.
   * 锚点种类，默认圆点 `dot`
   */
  type?: 'dot';
  /**
   * Percentage positioning of the anchor point relative to the node.
   * 锚点相对于节点的百分比定位(0-1)
   */
  position: number[];
  /**
   * Absolute offsets of the anchor point(px).
   * 锚点相对于节点的绝对偏移(px)
   */
  offsets?: number[];
  /**
   * Absolute offsets of the end point(px).
   * 锚点连接点的绝对偏移(px)
   */
  linkOffsets?: number[];
  /**
   * Size of the anchor.
   * 锚点大小
   */
  size?: number;
  /**
   * Visibility of the anchor.
   * 锚点可见性
   */
  visible?: boolean;
  /**
   * Index of the anchor.
   * 锚点序列
   */
  index?: number;
  /**
   * The way to display the anchor.
   * 锚点展示方式
   */
  show?: 'always' | 'hover';
  /**
   * The identifier of the anchor point being snapped.
   * 锚点被吸附标识
   */
  magnet?: boolean;
  /**
   * Customize the anchor styles.
   * 自定义锚点样式
   */
  setStyles?: (nodeData: any) => Record<string, unknown>;
  /**
   * Click handler of the anchor.
   * 锚点点击回调
   */
  onClick?: (e: ShapeEvent, nodeData: any) => void;
  /**
   * Mouse enter handler of the anchor.
   * 鼠标移入锚点回调
   */
  onMouseEnter?: (e: ShapeEvent, nodeData: any) => void;
  /**
   * Mouse leave handler of the anchor.
   * 鼠标移出锚点回调
   */
  onMouseLeave?: (e: ShapeEvent, nodeData: any) => void;
};